+++ easy_text/lib/easy_text.dart
@@ -3,4 +3,5 @@ library;
 export 'attributes.dart';
 export 'src/core/easy_text.dart';
 export 'src/core/easy_attribute_styles.dart';
+export 'src/core/changes/easy_text_change.dart';
 export 'src/extensions/runes_positioning.dart';